+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changed
 ^^^^^^^
 * Support ``os.PathLike`` in ``dump_open`` & ``dump_open_append``.
 * Support ``os.PathLike`` in ``open_offline``.
+* ``inject`` and ``sendpacket`` now take an object implementing the buffer protocol.
 
 v0.1.0b1 (2021-10-26)
 ---------------------